+++ template/typings/global/index.d.ts
@@ -1,14 +1,4 @@
-import { LogLevel } from 'nstarter-core/src/constants';
-
 declare global {
-    interface Error {
-        readonly code?: number | string;
-        readonly level?: LogLevel;
-        readonly meta?: any;
-        readonly httpCode?: number;
-        readonly isCustomError?: boolean;
-    }
-
     interface Constructor<T = any> {
         new(...args: any[]): T;
     }