+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Changelog
 
-## 0.7.0
+## 1.0.0
 
 * 依赖升级
   - 升级 node.js -> 14.18.x
   - 升级 typescript -> 4.4.x
+
+* 组件升级
   - 使用 axios 替换 request
+  - 升级 eslint-config-nstarter -> 2.0.x，对应变更 eslint 相关依赖
+  - 升级 socket.io -> 4.4.0
+  - 使用 @socket.io/redis-adapter 取代 socket.io-redis
+  - 使用 @socket.io/redis-emitter 取代 socket.io-emitter
+
+* 工具升级
+  - 升级 rimraf -> 3.0.2
+  - 升级 ts-node -> 10.4.0
+
 * 修复
   - 拆分 grpc server/client 组件的模板化配置